gc: reload js_closure_get_capture_bits results across a collection point (#7725)
The last 2 unrooted:capture hits gc-root-dominance-statepoints' budget
was carrying: js_closure_get_capture_bits(ptr, idx)'s return value is
never re-entered into the RS4GC-tracked domain or a temp root, unlike
%this_closure itself (already protected via current_closure_slot,
#7055). root_reload.rs's Facts now recognizes a capture-bits GET call
whose ptr operand already belongs to a reloadable recipe as an
extension of that recipe -- the same treatment string-handle globals
and shadow slots already get -- with a synthetic per-index store
side-condition so a same-index SET still suppresses the reload.

Native corpus: 2 -> 0 unrooted hazards, stale held at 0.
gc-root-dominance-statepoints' --max-unrooted is deleted (the flag
already defaults to 0) rather than set to 0, so a future hit can't be
silently absorbed by a stale budget.

root_reload.rs crossed the 2,000-line file-size cap gaining this; its
test module moved to a sibling root_reload_tests.rs (the existing
linker.rs/type_analysis.rs #[path] idiom), which is also where the
four new capture-bits tests landed.
