+++ changelog.d/7725-capture-bits-reload.md
@@ -1,0 +1,31 @@
+### gc-root-dominance-statepoints: the last 2 `unrooted:capture` hits are fixed, `--max-unrooted` deleted
+
+Closes #7725, the follow-up #7724 split out rather than rush: `js_closure_get_capture_bits(ptr, idx)` reads a captured value out of the closure's own capture array — a location the collector rewrites when it relocates the closure — but unlike `%this_closure` itself (already re-entered into the `ptr addrspace(1)` tracked domain, #7055), the call's *return value* was never re-entered into either that domain or a temp root. Two corpus hits, both `read once, held long, across an allocating call`:
+
+- `test_gap_class_expr_dynamic_parent_ctor::__closure_21` — the synthesized dynamic-parent implicit constructor reads a captured base-class reference at the top of the closure and uses it ~60 lines later as `js_new_function_construct`'s callee, across `js_object_alloc_class_inline_keys` and the class's own user constructor body.
+- `test_gap_computed_key_method_nested_this::__closure_9` — a captured numeric local held across `js_number_coerce` (which can run a user `Symbol.toPrimitive`).
+
+#### Confirming the framing before implementing
+
+#7724's triage read this as a gap in the *repair machinery*: `root_reload.rs`'s `Facts` only modeled a *load* as a reloadable source, not a *call*, even though re-calling `js_closure_get_capture_bits` is sound under the identical store side-condition (a `js_closure_set_capture_bits` to the same index) the pass already tracks for shadow slots and string-handle globals. The alternative — rooting every capture read at its own call site via the existing `TempRootPool` (`rooting/temp_root.rs`) — was rejected: it would pay a store + shadow-slot bind on *every* capture read, including the overwhelming majority never held across a collection point, where `root_reload.rs`'s whole design point is to cost nothing where the reload is unnecessary (LLVM CSEs a redundant reload away; an actually-needed one is opaque and stays). The generic-read call sites (`literals_vars.rs`, `closure.rs`, `array_push.rs`, `instance_misc1.rs`, `lower_array_method.rs`, `expr/mod.rs`) also have no way to know at emission time whether *this* read will cross a later collection — that's exactly the question `root_reload.rs`'s post-hoc CFG walk answers cheaply. So the `Facts`-as-call change was the right layer, not over-engineering.
+
+#### The fix
+
+`js_closure_get_capture_bits` calls whose `ptr` operand already belongs to a reloadable recipe (the closure-ptr shadow-slot load chain #7055 protects) now extend that recipe exactly like a transparent `and`/`bitcast` step — `Facts` gained `capture_get_key`, set when the call's index argument is a compile-time literal (every real emission site formats one directly). A reload point re-materializes the *whole* chain: a fresh slot load, a fresh mask, a fresh call to `js_closure_get_capture_bits` with the same index, and any further cast — not just the load.
+
+The store side-condition needed its own key, not the inherited closure-ptr-slot one (that slot is written exactly once, at closure entry, so checking it would make the invalidation vacuous): a `js_closure_get_capture_bits` recipe's `root_ptr` switches to a synthetic `$closure_capture:<idx>` key from that call onward, and `js_closure_set_capture_bits` populates the existing `stores_to` field with the same key — reusing the store-invalidation machinery a plain `store` already has, rather than adding a parallel mechanism. That required widening the reload pass's grouping key from `recipe[0]` alone to `(recipe[0], root_ptr)`, since a chain's `root_ptr` had never changed mid-derivation before this — every pre-existing chain has one `root_ptr` throughout, so the wider key is additive and changes no existing grouping.
+
+`raw_facts` (the `invoke`/`try` rendering path) gets the same `stores_to` detection for the set half, but the get half is gated on `!is_invoke`: `materialize` re-emits a recipe step as a plain mid-block instruction, and an `invoke` is a terminator — cloning one in would be invalid IR and would re-target the original call's `%cont`/`%lpad` labels from a second predecessor. In practice this call is never seen here except as an invoke (every real emission goes through the typed `LlInst::Call` form otherwise), so the guard is defensive rather than load-bearing today.
+
+#### Verified
+
+- `python3 scripts/gc_root_dominance_check.py ir-corpus-native --statepoints --moving-only --max-unrooted 999 --max-stale 999 -v`: **2 → 0** `unrooted`, `stale` held at 0, no other hazard class changed (native corpus, 131/131 sources compiled, 0 skipped both before and after).
+- The full CI-shaped invocation (`--min-statepoints 15000 --min-live-bundles 8000 --min-relocates 20000 --max-unrooted 0 --max-stale 0 --seeded-violations 40`) passes; the 40 planted violations are still all caught, proving the checker itself wasn't accidentally neutered.
+- `--lowering shadow` corpus (131/131 compiled): the dominance check, `--unrooted-allocas --moving-only`, and `--stale-registers --moving-only` (6 pre-existing hits, unrelated to capture-bits, well under the 39 budget) all stay clean.
+- `python3 scripts/gc_root_dominance_check.py --self-test`: clean, including the `phi_safe_edge`/`phi_hazard_edge` fixtures #7724 added — confirms this change didn't touch the checker itself.
+- `cargo test -p perry-codegen --lib`: 809 passed (805 on `origin/main` + 4 new capture-bits reload tests), 0 failed.
+- `cargo test -p perry-codegen --no-fail-fast` (the full suite, including the `crates/*/tests/*.rs` integration binaries CI does not run per-PR, #5960): the same 6 pre-existing failures across the same 4 targets (`large_object_barriers`, `native_proof_buffer_views`, `native_proof_regressions`, `typed_shape_descriptors`) reproduce byte-identically on a fresh `origin/main` checkout — zero new failures.
+
+`gc-root-dominance-statepoints`'s `--max-unrooted` is **deleted** rather than set to `0` (the flag already defaults to 0): a budget above reality silently absorbs the next real hazard, which is the exact defect #7706 fixed for a different gate on this file.
+
+`crates/perry-codegen/src/root_reload.rs` crossed the 2,000-line cap while gaining this; its `#[cfg(test)] mod tests` moved to a sibling `root_reload_tests.rs` (same `--path` idiom `linker.rs`/`type_analysis.rs` already use), which is also where the four new tests landed.
